booking_site/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
pydantic==2.12.5

starlette==0.52.1
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
python-multipart==0.0.22

httpx==0.28.1
httpcore==1.0.9

anyio==4.12.1
typing_extensions==4.15.0
annotated-types==0.7.0